backend/emosense/requirements.txt
absl-py==2.1.0
altair==5.2.0
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
autopep8==2.0.2
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.2
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.3
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.2
django-allauth==0.61.1
django-cleanup==8.0.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-jwt==1.0.1
executing==2.0.1
facenet-pytorch==2.5.3
fastapi==0.109.0
filelock==3.13.1
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.47.0
fsspec==2023.12.2
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-oauth2-tool==0.0.3
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.62.1
h11==0.14.0
h5py==3.11.0
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
ipython==8.20.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.2.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
mysql-connector==2.2.9
mysqlclient==2.2.4
namex==0.0.8
networkx==3.2.1
numpy==1.26.3
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==4.25.3
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
selenium==4.16.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.35.1
streamlit==1.31.1
sympy==1.12
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.3.0
toml==0.10.2
toolz==0.12.1
torch==2.2.0
torchvision==0.17.0
tornado==6.4
traitlets==5.14.1
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.0.3
uvicorn==0.26.0
validators==0.22.0
virtualenv==20.25.0
watchdog==4.0.0
wcwidth==0.2.13
webdriver-manager==4.0.1
Werkzeug==3.0.1
wrapt==1.16.0
wsproto==1.2.0
xgboost==1.4.2
zipp==3.17.0
